requirements.txt
requests 
python-chess